+++ changelog.md
@@ -2,8 +2,15 @@
 
 ## Version 3.2.24
 
+- Added feature allowing folders of items to be dropped onto item piles to fill item pile with all items contained in folder
+- Added feature allowing rolltables and folders of rolltables to dropped onto item piles - each rolltable will be rolled once, and the result will be added to the item pile
+- Added support for rolltables containing text entries like:
+  - `3gp` - 3 GP will be added to item pile
+  - `1d6gp` - 1-6 GP will be added to the item pile
+  - `2d6gp 50sp` - 2-12 GP and 50 SP will be added to the item pile
 - Fixed support for rounding up when the primary currency was that the lowest exchange rate of all currencies
 - Fixed currency editor sorting currencies by exchange rate on keydown, rather than on change
+- Upgraded `API#getPaymentDataFromString` to a loud deprecation
 
 ## Version 3.2.23